+++ src/lib/types/index.ts
@@ -2,4 +2,4 @@
 
 export * from "@/lib/types/course";
 export * from "@/lib/types/filters";
-export * from "@/lib/types/planner";
+export * from "@/lib/types/semester";
